Move EditorManager to ace.dart
